+++ CHANGELOG.md
@@ -9,6 +9,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 - Added `test.sh` for a "one-click" verification of the client-server setup.
+
+### Changed
+- Pre-configured `test.sh` with a default host path to simplify usage.
 - Renamed `test-project` to `example-project` and included it in the repository to serve as a ready-to-use test case.
 
 ## [0.4.0] - 2026-02-09